Fixed warnings

+++ ph-as4-lib/src/main/java/com/helger/as4/model/pmode/leg/PModeLegBusinessInformationMicroTypeConverter.java
@@ -47,7 +47,7 @@ public class PModeLegBusinessInformationMicroTypeConverter extends
                                                                                                   sNamespaceURI,
                                                                                                   ELEMENT_PAYLOAD_PROFILE)));
     if (aValue.getPayloadProfileMaxKB () != null)
-      ret.setAttribute (ATTR_PAYLOAD_PROFILE_MAX_KB, aValue.getPayloadProfileMaxKB ());
+      ret.setAttribute (ATTR_PAYLOAD_PROFILE_MAX_KB, aValue.getPayloadProfileMaxKB ().longValue ());
     ret.setAttribute (ATTR_MPCID, aValue.getMPCID ());
     return ret;
   }
